Centralizing mongodb connection
